FIX: migrations for mysql

+++ database/migrations/2020_09_19_200150_create_relations_tables.php
@@ -14,8 +14,8 @@ class CreateRelationsTables extends Migration
   public function up()
   {
     Schema::create('character_comic', function (Blueprint $table) {
-      $table->integer('character_id')->unsigned();
-      $table->integer('comic_id')->unsigned();
+      $table->bigInteger('character_id')->unsigned();
+      $table->bigInteger('comic_id')->unsigned();
       $table->foreign('character_id')->references('id')->on('characters')
         ->onDelete('cascade');
       $table->foreign('comic_id')->references('id')->on('comics')
@@ -23,8 +23,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('character_story', function (Blueprint $table) {
-      $table->integer('character_id')->unsigned();
-      $table->integer('story_id')->unsigned();
+      $table->bigInteger('character_id')->unsigned();
+      $table->bigInteger('story_id')->unsigned();
       $table->foreign('character_id')->references('id')->on('characters')
         ->onDelete('cascade');
       $table->foreign('story_id')->references('id')->on('stories')
@@ -32,8 +32,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('character_event', function (Blueprint $table) {
-      $table->integer('character_id')->unsigned();
-      $table->integer('event_id')->unsigned();
+      $table->bigInteger('character_id')->unsigned();
+      $table->bigInteger('event_id')->unsigned();
       $table->foreign('character_id')->references('id')->on('characters')
         ->onDelete('cascade');
       $table->foreign('event_id')->references('id')->on('events')
@@ -41,8 +41,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('character_serie', function (Blueprint $table) {
-      $table->integer('character_id')->unsigned();
-      $table->integer('serie_id')->unsigned();
+      $table->bigInteger('character_id')->unsigned();
+      $table->bigInteger('serie_id')->unsigned();
       $table->foreign('character_id')->references('id')->on('characters')
         ->onDelete('cascade');
       $table->foreign('serie_id')->references('id')->on('series')
@@ -50,8 +50,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('comic_story', function (Blueprint $table) {
-      $table->integer('comic_id')->unsigned();
-      $table->integer('story_id')->unsigned();
+      $table->bigInteger('comic_id')->unsigned();
+      $table->bigInteger('story_id')->unsigned();
       $table->foreign('comic_id')->references('id')->on('comics')
         ->onDelete('cascade');
       $table->foreign('story_id')->references('id')->on('stories')
@@ -59,8 +59,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('comic_event', function (Blueprint $table) {
-      $table->integer('comic_id')->unsigned();
-      $table->integer('event_id')->unsigned();
+      $table->bigInteger('comic_id')->unsigned();
+      $table->bigInteger('event_id')->unsigned();
       $table->foreign('comic_id')->references('id')->on('comics')
         ->onDelete('cascade');
       $table->foreign('event_id')->references('id')->on('events')
@@ -68,8 +68,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('comic_serie', function (Blueprint $table) {
-      $table->integer('comic_id')->unsigned();
-      $table->integer('serie_id')->unsigned();
+      $table->bigInteger('comic_id')->unsigned();
+      $table->bigInteger('serie_id')->unsigned();
       $table->foreign('comic_id')->references('id')->on('comics')
         ->onDelete('cascade');
       $table->foreign('serie_id')->references('id')->on('series')
@@ -77,8 +77,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('story_event', function (Blueprint $table) {
-      $table->integer('story_id')->unsigned();
-      $table->integer('event_id')->unsigned();
+      $table->bigInteger('story_id')->unsigned();
+      $table->bigInteger('event_id')->unsigned();
       $table->foreign('story_id')->references('id')->on('stories')
         ->onDelete('cascade');
       $table->foreign('event_id')->references('id')->on('events')
@@ -86,8 +86,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('story_serie', function (Blueprint $table) {
-      $table->integer('story_id')->unsigned();
-      $table->integer('serie_id')->unsigned();
+      $table->bigInteger('story_id')->unsigned();
+      $table->bigInteger('serie_id')->unsigned();
       $table->foreign('story_id')->references('id')->on('stories')
         ->onDelete('cascade');
       $table->foreign('serie_id')->references('id')->on('series')
@@ -95,8 +95,8 @@ class CreateRelationsTables extends Migration
     });
 
     Schema::create('event_serie', function (Blueprint $table) {
-      $table->integer('event_id')->unsigned();
-      $table->integer('serie_id')->unsigned();
+      $table->bigInteger('event_id')->unsigned();
+      $table->bigInteger('serie_id')->unsigned();
       $table->foreign('event_id')->references('id')->on('events')
         ->onDelete('cascade');
       $table->foreign('serie_id')->references('id')->on('series')
